+++ code/wave_models/src/Airy.cpp
@@ -32,7 +32,7 @@ generate_random_phase(boost::random::uniform_real_distribution<double>(0,2*PI))
 
 Airy::Airy(const DiscreteDirectionalWaveSpectrum& spectrum_, const int random_number_generator_seed) : WaveModel(spectrum_),
 phase(std::vector<std::vector<double> >()),
-rng(boost::mt19937(random_number_generator_seed)),
+rng(boost::mt19937((unsigned int)random_number_generator_seed)),
 generate_random_phase(boost::random::uniform_real_distribution<double>(0,2*PI))
 {
     for (size_t i = 0 ; i < spectrum.omega.size() ; ++i)
